Avoid invalid UTF-8 exception in invalid SQL statement handler.
The mysql connection adapter tries to provide a helpful message for a
certain class of invalid SQL statement exceptions. However, if the
exception has invalid UTF-8 characters, the handler raises a UTF-8
exception that is not very helpful for debugging.

This commit makes the exception handler pass the original exception
through.